add `brave_payments_enabled` flag and disable on windows for now
